+++ packages/google_sign_in/google_sign_in_android/lib/google_sign_in_android.dart
@@ -30,11 +30,22 @@ class GoogleSignInAndroid extends GoogleSignInPlatform {
     String? hostedDomain,
     String? clientId,
   }) {
+    return initWithParams(SignInInitParameters(
+      signInOption: signInOption,
+      scopes: scopes,
+      hostedDomain: hostedDomain,
+      clientId: clientId,
+    ));
+  }
+
+  @override
+  Future<void> initWithParams(SignInInitParameters params) {
     return channel.invokeMethod<void>('init', <String, dynamic>{
-      'signInOption': signInOption.toString(),
-      'scopes': scopes,
-      'hostedDomain': hostedDomain,
-      'clientId': clientId,
+      'signInOption': params.signInOption.toString(),
+      'scopes': params.scopes,
+      'hostedDomain': params.hostedDomain,
+      'clientId': params.clientId,
+      'forceCodeForRefreshToken': params.forceCodeForRefreshToken,
     });
   }