+++ packages/google_sign_in/google_sign_in_web/lib/google_sign_in_web.dart
@@ -71,8 +71,18 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
     SignInOption signInOption = SignInOption.standard,
     String? hostedDomain,
     String? clientId,
-  }) async {
-    final String? appClientId = clientId ?? _autoDetectedClientId;
+  }) {
+    return initWithParams(SignInInitParameters(
+      signInOption: signInOption,
+      scopes: scopes,
+      hostedDomain: hostedDomain,
+      clientId: clientId,
+    ));
+  }
+
+  @override
+  Future<void> initWithParams(SignInInitParameters params) async {
+    final String? appClientId = params.clientId ?? _autoDetectedClientId;
     assert(
         appClientId != null,
         'ClientID not set. Either set it on a '
@@ -80,7 +90,7 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
         ' or pass clientId when calling init()');
 
     assert(
-        !scopes.any((String scope) => scope.contains(' ')),
+        !params.scopes.any((String scope) => scope.contains(' ')),
         "OAuth 2.0 Scopes for Google APIs can't contain spaces. "
         'Check https://developers.google.com/identity/protocols/googlescopes '
         'for a list of valid OAuth 2.0 scopes.');
@@ -88,9 +98,9 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
     await _isGapiInitialized;
 
     final auth2.GoogleAuth auth = auth2.init(auth2.ClientConfig(
-      hosted_domain: hostedDomain,
+      hosted_domain: params.hostedDomain,
       // The js lib wants a space-separated list of values
-      scope: scopes.join(' '),
+      scope: params.scopes.join(' '),
       client_id: appClientId!,
       plugin_name: 'dart-google_sign_in_web',
     ));
@@ -118,22 +128,6 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
     }));
 
     return _isAuthInitialized;
-  }
-
-  @override
-  Future<void> initWithForceCodeForRefreshToken({
-    List<String> scopes = const <String>[],
-    SignInOption signInOption = SignInOption.standard,
-    String? hostedDomain,
-    String? clientId,
-    bool forceCodeForRefreshToken = false,
-  }) async {
-    await init(
-      scopes: scopes,
-      signInOption: signInOption,
-      hostedDomain: hostedDomain,
-      clientId: clientId,
-    );
   }
 
   @override